requirements.txt
beautifulsoup4==4.12.2
et-xmlfile==1.1.0
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
soupsieve==2.5
tzdata==2023.3